Move github and travis api calls out into utilities
